requirements.txt
python-docx>=1.1.0
pypdf>=3.0.0


